travel_assistant/requirements.txt
art==5.9
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
Flask==3.1.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
pyfiglet==1.0.4
Pygments==2.19.2
python-dotenv==1.1.1
requests==2.32.5
rich==14.2.0
urllib3==2.5.0
Werkzeug==3.1.3
